search: factor out zoekt query logic in query_converter.go
