+++ internal/search/query_converter.go
@@ -1,7 +1,6 @@
 package search
 
 import (
-	"regexp/syntax" //nolint:depguard // zoekt requires this pkg
 	"strings"
 	"time"
 
@@ -12,10 +11,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/search/limits"
 	"github.com/sourcegraph/sourcegraph/internal/search/query"
-	"github.com/sourcegraph/sourcegraph/internal/search/result"
-	"github.com/sourcegraph/sourcegraph/lib/errors"
-
-	zoekt "github.com/google/zoekt/query"
 )
 
 // UnionRegExps separates values with a | operator to create a string
@@ -67,14 +62,6 @@ func LangToFileRegexp(lang string) string {
 	return UnionRegExps(patterns)
 }
 
-func mapSlice(values []string, f func(string) string) []string {
-	result := make([]string, len(values))
-	for i, v := range values {
-		result[i] = f(v)
-	}
-	return result
-}
-
 type Protocol int
 
 const (
@@ -96,191 +83,4 @@ func TimeoutDuration(b query.Basic) time.Duration {
 		d = maxTimeout
 	}
 	return d
-}
-
-func fileRe(pattern string, queryIsCaseSensitive bool) (zoekt.Q, error) {
-	return parseRe(pattern, true, false, queryIsCaseSensitive)
-}
-
-func noOpAnyChar(re *syntax.Regexp) {
-	if re.Op == syntax.OpAnyChar {
-		re.Op = syntax.OpAnyCharNotNL
-	}
-	for _, s := range re.Sub {
-		noOpAnyChar(s)
-	}
-}
-
-func parseRe(pattern string, filenameOnly bool, contentOnly bool, queryIsCaseSensitive bool) (zoekt.Q, error) {
-	// these are the flags used by zoekt, which differ to searcher.
-	re, err := syntax.Parse(pattern, syntax.ClassNL|syntax.PerlX|syntax.UnicodeGroups)
-	if err != nil {
-		return nil, err
-	}
-	noOpAnyChar(re)
-	// zoekt decides to use its literal optimization at the query parser
-	// level, so we check if our regex can just be a literal.
-	if re.Op == syntax.OpLiteral {
-		return &zoekt.Substring{
-			Pattern:       string(re.Rune),
-			CaseSensitive: queryIsCaseSensitive,
-			Content:       contentOnly,
-			FileName:      filenameOnly,
-		}, nil
-	}
-	return &zoekt.Regexp{
-		Regexp:        re,
-		CaseSensitive: queryIsCaseSensitive,
-		Content:       contentOnly,
-		FileName:      filenameOnly,
-	}, nil
-}
-
-func toZoektPattern(expression query.Node, isCaseSensitive, patternMatchesContent, patternMatchesPath bool) (zoekt.Q, error) {
-	var fold func(node query.Node) (zoekt.Q, error)
-	fold = func(node query.Node) (zoekt.Q, error) {
-		switch n := node.(type) {
-		case query.Operator:
-			children := make([]zoekt.Q, 0, len(n.Operands))
-			for _, op := range n.Operands {
-				child, err := fold(op)
-				if err != nil {
-					return nil, err
-				}
-				children = append(children, child)
-			}
-			switch n.Kind {
-			case query.Or:
-				return &zoekt.Or{Children: children}, nil
-			case query.And:
-				return &zoekt.And{Children: children}, nil
-			default:
-				// unreachable
-				return nil, errors.Errorf("broken invariant: don't know what to do with node %T in toZoektPattern", node)
-			}
-		case query.Pattern:
-			var q zoekt.Q
-			var err error
-			fileNameOnly := patternMatchesPath && !patternMatchesContent
-			contentOnly := !patternMatchesPath && patternMatchesContent
-
-			pattern := n.Value
-			if n.Annotation.Labels.IsSet(query.Literal) {
-				pattern = regexp.QuoteMeta(pattern)
-			}
-
-			q, err = parseRe(pattern, fileNameOnly, contentOnly, isCaseSensitive)
-			if err != nil {
-				return nil, err
-			}
-
-			if n.Negated {
-				q = &zoekt.Not{Child: q}
-			}
-			return q, nil
-		}
-		// unreachable
-		return nil, errors.Errorf("broken invariant: don't know what to do with node %T in toZoektPattern", node)
-	}
-
-	q, err := fold(expression)
-	if err != nil {
-		return nil, err
-	}
-
-	return q, nil
-}
-
-func QueryToZoektQuery(b query.Basic, resultTypes result.Types, feat *Features, typ IndexedRequestType) (q zoekt.Q, err error) {
-	isCaseSensitive := b.IsCaseSensitive()
-
-	if b.Pattern != nil {
-		q, err = toZoektPattern(
-			b.Pattern,
-			isCaseSensitive,
-			resultTypes.Has(result.TypeFile),
-			resultTypes.Has(result.TypePath),
-		)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Handle file: and -file: filters.
-	filesInclude, filesExclude := b.IncludeExcludeValues(query.FieldFile)
-	// Handle lang: and -lang: filters.
-	langInclude, langExclude := b.IncludeExcludeValues(query.FieldLang)
-	filesInclude = append(filesInclude, mapSlice(langInclude, LangToFileRegexp)...)
-	filesExclude = append(filesExclude, mapSlice(langExclude, LangToFileRegexp)...)
-	filesReposMustInclude, filesReposMustExclude := b.IncludeExcludeValues(query.FieldRepoHasFile)
-
-	if typ == SymbolRequest && q != nil {
-		// Tell zoekt q must match on symbols
-		q = &zoekt.Symbol{
-			Expr: q,
-		}
-	}
-
-	var and []zoekt.Q
-	if q != nil {
-		and = append(and, q)
-	}
-
-	// zoekt also uses regular expressions for file paths
-	// TODO PathPatternsAreCaseSensitive
-	// TODO whitespace in file path patterns?
-	for _, i := range filesInclude {
-		q, err := fileRe(i, isCaseSensitive)
-		if err != nil {
-			return nil, err
-		}
-		and = append(and, q)
-	}
-	if len(filesExclude) > 0 {
-		q, err := fileRe(UnionRegExps(filesExclude), isCaseSensitive)
-		if err != nil {
-			return nil, err
-		}
-		and = append(and, &zoekt.Not{Child: q})
-	}
-
-	// For conditionals that happen on a repo we can use type:repo queries. eg
-	// (type:repo file:foo) (type:repo file:bar) will match all repos which
-	// contain a filename matching "foo" and a filename matchinb "bar".
-	//
-	// Note: (type:repo file:foo file:bar) will only find repos with a
-	// filename containing both "foo" and "bar".
-	for _, i := range filesReposMustInclude {
-		q, err := fileRe(i, isCaseSensitive)
-		if err != nil {
-			return nil, err
-		}
-		and = append(and, &zoekt.Type{Type: zoekt.TypeRepo, Child: q})
-	}
-	for _, i := range filesReposMustExclude {
-		q, err := fileRe(i, isCaseSensitive)
-		if err != nil {
-			return nil, err
-		}
-		and = append(and, &zoekt.Not{Child: &zoekt.Type{Type: zoekt.TypeRepo, Child: q}})
-	}
-
-	// Languages are already partially expressed with IncludePatterns, but Zoekt creates
-	// more precise language metadata based on file contents analyzed by go-enry, so it's
-	// useful to pass lang: queries down.
-	//
-	// Currently, negated lang queries create filename-based ExcludePatterns that cannot be
-	// corrected by the more precise language metadata. If this is a problem, indexed search
-	// queries should have a special query converter that produces *only* Language predicates
-	// instead of filepatterns.
-	if len(langInclude) > 0 && feat.ContentBasedLangFilters {
-		or := &zoekt.Or{}
-		for _, lang := range langInclude {
-			lang, _ = enry.GetLanguageByAlias(lang) // Invariant: lang is valid.
-			or.Children = append(or.Children, &zoekt.Language{Language: lang})
-		}
-		and = append(and, or)
-	}
-
-	return zoekt.Simplify(zoekt.NewAnd(and...)), nil
 }
